bug-fix/fix-eventType-mapping-and-more

+++ src/main/kotlin/com/babytrackr/analytics/controller/model/FeedTotals.kt
@@ -2,5 +2,5 @@ package com.babytrackr.analytics.controller.model
 
 data class FeedTotals(
     val totalFeedings: Int,
-    val totalOunces: Int,
+    val totalOunces: Double,
 )
