+++ src/main/kotlin/com/babytrackr/analytics/controller/model/response/FeedBreakdown.kt
@@ -3,5 +3,5 @@ package com.babytrackr.analytics.controller.model.response
 data class FeedBreakdown(
     val label: String,
     val totalFeedings: Int,
-    val totalOunces: Int
+    val totalOunces: Double
 )
